Extract base from browser test to be used in other tests.
