+++ test/browser/shared/wizard.js
@@ -5,12 +5,11 @@ module.exports = function(settings) {
 
     return async function(runner, screens) {
         const container = await runner.$('#app')
-        if (screens) await container.screenshot({path: path.join(settings.SCREENS_DIR, `${brand.tests.step(runner)}wizard-welcome.png`)})
+        if (screens) await brand.tests.screenshot(container, runner, 'wizard-welcome')
 
         await runner.click('.test-wizard-welcome-next')
-
         await runner.waitFor('.component-wizard-telemetry')
-        if (screens) await container.screenshot({path: path.join(settings.SCREENS_DIR, `${brand.tests.step(runner)}wizard-telemetry.png`)})
+        if (screens) await brand.tests.screenshot(container, runner, 'wizard-telemetry')
 
         await runner.click('.test-wizard-telemetry-yes')
 
@@ -23,12 +22,12 @@ module.exports = function(settings) {
             await runner.click('input[id="webrtc_account"]')
             await runner.click(`.filtered-options #option-${brand.tests[runner._name].id}`)
 
-            if (screens) await container.screenshot({path: path.join(settings.SCREENS_DIR, `${brand.tests.step(runner)}wizard-account.png`)})
+            if (screens) await brand.tests.screenshot(container, runner, 'wizard-account')
             await runner.click('.test-wizard-account-next')
         }
 
         await runner.waitFor('.component-wizard-mic-permission')
-        if (screens) await container.screenshot({path: path.join(settings.SCREENS_DIR, `${brand.tests.step(runner)}wizard-mic-permission.png`)})
+        if (screens) await brand.tests.screenshot(container, runner, 'wizard-mic-permission')
         await runner.click('.test-wizard-mic-permission-next')
 
         await runner.waitFor('.component-wizard-devices')
@@ -40,7 +39,7 @@ module.exports = function(settings) {
             runner.$$('#sounds_device option'),
         ])
 
-        if (screens) await container.screenshot({path: path.join(settings.SCREENS_DIR, `${brand.tests.step(runner)}wizard-devices.png`)})
+        if (screens) await brand.tests.screenshot(container, runner, 'wizard-devices')
         await runner.click('.test-wizard-devices-next')
 
         await runner.waitFor('.notification')
